Remove GA tracking. IANAL.
I'm not sure whether GA tracking is legal without informing the user
(very likely not). Probably I must provide some cookie informational
banner. But I do not know how for now. This project is rather about the
fun, so until I want to invest in the lawyer-how-to, I'm rather removing
the tracking.